requirements.txt
numpy==2.0.2
pandas==2.2.2
scikit-learn==1.6.1
xgboost==3.1.2
lightgbm==4.6.0
matplotlib==3.10.0
seaborn==0.13.2
pyyaml==6.0.1